+++ protected/humhub/docs/CHANGELOG.md
@@ -6,6 +6,7 @@ HumHub Change Log
 -------------------------
 
 - Fix #3647: GroupID field is not translatable in Registration Form
+- Fix #3655: Fix CommentLink widget Comment class usage conflict
 
 
 1.3.15  (August 22, 2019)